提取SHX对话框抑制核心为共享模块
将AutoCAD“缺少SHX文件”对话框抑制器核心算法抽取为AwsHideableDialogPatcherCore.cs，放入Shared目录，插件与部署器通过Link方式共用。插件侧与部署器侧相关类重构为参数收集与日志转发的薄包装，实际逻辑全部委托共享核心。核心方法参数化brand/version/registryBasePath，彻底去除对PlatformManager和ICadPlatform依赖，实现零重复。同步调整解决方案结构与项目引用，完善注释说明复用边界。
